DC-09: Address code review — io.LimitReader consistency, test coverage gaps
- Normalize io.LimitReader call to use int64() cast in doRequest,
  matching fetchFullText and PubMed plugin patterns
- Add test for unsupported format (FormatXML → ErrFormatUnsupported)
- Add test for full text skip on non-OA article (IsOpenAccess != "Y")
- Add test for empty full text body (200 OK with zero bytes)

+++ internal/rtv.plugin.europmc.go
@@ -516,7 +516,7 @@ func (p *EuropePMCPlugin) doRequest(ctx context.Context, reqURL string, target a
 		return fmt.Errorf("%w: "+emcHTTPStatusErrFmt, ErrEuropePMCHTTPRequest, resp.StatusCode)
 	}
 
-	limitedBody := io.LimitReader(resp.Body, emcMaxResponseBytes)
+	limitedBody := io.LimitReader(resp.Body, int64(emcMaxResponseBytes))
 	if err := json.NewDecoder(limitedBody).Decode(target); err != nil {
 		return fmt.Errorf("%w: %w", ErrEuropePMCJSONParse, err)
 	}
